adapted to new web version a browsers
Toyota has recently changed its page to a new version so the extension was not working anymore.

With this new version it will work again, and I also solve a problem withe the manifest version where a some users where experiencing errors due to an old manifest version was not allowed for their browsers.